pyproject.toml
[project]
name = "openai-agents-ext"
version = "0.1.0"
description = "An extension library for the OpenAI Agents SDK, providing enhanced streaming capabilities."
readme = "README.md"
authors = [
    {name = "RussellLuo", email = "luopeng.he@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents_ext"]
